On branch development Your branch is up to date with 'origin/development'.
Changes to be committed:
	modified:   src/chem_pathway_analysis/i__user_model/convert_concentration_file.py
	modified:   src/chem_pathway_analysis/o__cpap_output/output.py
	modified:   src/chem_pathway_analysis/p__data_management/data_check.py
	modified:   src/chem_pathway_analysis/p__data_management/data_tools.py
	modified:   src/chem_pathway_analysis/p__data_management/data_update.py
	modified:   src/chem_pathway_analysis/p__pathways_analysis/branching_points.py
	modified:   src/chem_pathway_analysis/p__pathways_analysis/main_loop.py
	modified:   src/chem_pathway_analysis/p__sub_pathways/subpathway_analysis.py
	modified:   test_cpa.py

The code is in a pretty good shape. BUG corrected in checking if a pathways is present.
Using the couple {reaction index:reaction multiplicity} and not juste the reaction index
NONETHELESS: Issue with test Uranus with f_min=1.0e-20. We have an issue with the Sub Pathways.
It looks like the building of a P with its elementary SubP is going wrong. I'm not sure the
definition of an elementary Sub Pathway is well done. It looks like some Sub P cannot combine
and explain the Pathway. Look for the results of the linear problem and how it's updated.
Example : 1 P {1:1,2:1,3:1} with two Sub P sP1 = {1:1,2:3} and sP2 = {3:1}
        : We see that P != sP1 + sP2 and cannot be any combination of sP1 and sP2
        : Nonetheless, the operation set(sP1).issubset(P) = True, stuff like that
        : But the result of the linear probelm seems to be 1 D, meaning only one sP
        : It could indicate that there is actually not possible combination
        : So that the initial pathway is still not changed
        : But possibly the update is going wrong, and we choose sP1 instead